benchmarks/requirements-bench.txt
numpy>=1.24
matplotlib>=3.7
seaborn>=0.12
click>=8.1
pyzmq>=25.0
pynng>=0.7.2
